requirements.txt

diffusers
transformers
faiss-cpu
numpy
Pillow
moviepy
streamlit

opencv-python
insightface
accelerate